+++ placement-flutter/storyly_placement_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+# Release Notes
 
-* TODO: Describe initial release.
+### 1.2.1 (16.01.2026)
+* added storyly placement sdk
+* added storyly swipe card widget
+* added storyly banner widget
+* added storyly story bar widget
+* added storyly video feed widget
+* added storyly canvas widget